Moved the "inspect rune" logic into its own module and fixed some bugs in its output
